Ramenを別ファイルに分けた
