+++ ImageSaveSampleApp/Model/RamenModel.swift
@@ -8,14 +8,6 @@
 import SwiftUI
 import FirebaseStorage
 
-struct Ramen: Identifiable {
-    var id = UUID() // 自動生成される一意のID
-    var name: String
-    var shop: String
-    var rating: Int
-    var imageUrl: String? // Firebase Storageから取得した画像のURL（まだ画像がない場合はnil）
-}
-
 struct RamenModel {
     // firebaseをインスタンスしてリポジトリを使用する
     private var firebaseStorageRepository = FirebaseStorageRepository()
